requirements.txt
streamlit>=1.31.0
openai>=1.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
pandas>=2.1.0
pydantic>=2.5.0
python-dateutil>=2.8.0
# Async support
aiohttp>=3.9.0
# Structured logging
structlog>=23.2.0
# Timeout handling
httpx>=0.25.0  # For async HTTP with timeout support
# Dashboard visualizations
plotly>=5.18.0  # For interactive charts
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0